+++ packages/react-router-dom/CHANGELOG.md
@@ -1,5 +1,13 @@
 # `react-router-dom`
 
+## 6.14.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@6.14.1`
+  - `@remix-run/router@1.7.1`
+
 ## 6.14.0
 
 ### Minor Changes