+++ packages/react-router-native/CHANGELOG.md
@@ -1,5 +1,26 @@
 # `react-router-native`
 
+## 6.14.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@6.14.1`
+
+## 6.14.1-pre.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@6.14.1-pre.1`
+
+## 6.14.1-pre.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@6.14.1-pre.0`
+
 ## 6.14.0
 
 ### Patch Changes